requirements.txt
altgraph==0.16.1
aniso8601==6.0.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
macholib==1.11
MarkupSafe==1.1.1
pefile==2018.8.8
pycryptodomex==3.8.1
PyInstaller==3.4
pypiwin32==223
pytz==2019.1
pywin32==224
pywin32-ctypes==0.2.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
